♻️ refactor(server): isolate feishu channel handler
